# UnphuVacantesAPI/wwwroot/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "UnphuVacantesAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Empresas": {
      "get": {
        "tags": [
          "Empresas"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Empresa"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Empresa"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Empresa"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Empresas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empresa"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Empresa"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Empresa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Empresa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empresa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empresa"
                }
              }
            }
          }
        }
      }
    },
    "/api/Empresas/{id}": {
      "get": {
        "tags": [
          "Empresas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Empresa"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empresa"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empresa"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Empresas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empresa"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Empresa"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Empresa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Empresas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Postulaciones": {
      "get": {
        "tags": [
          "Postulaciones"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Postulacion"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Postulacion"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Postulacion"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Postulaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Postulacion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Postulacion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Postulacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Postulacion"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulacion"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulacion"
                }
              }
            }
          }
        }
      }
    },
    "/api/Postulaciones/{id}": {
      "get": {
        "tags": [
          "Postulaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Postulacion"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulacion"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulacion"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Postulaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Postulantes": {
      "get": {
        "tags": [
          "Postulantes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Postulante"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Postulante"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Postulante"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Postulantes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Postulante"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Postulante"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Postulante"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Postulante"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulante"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulante"
                }
              }
            }
          }
        }
      }
    },
    "/api/Postulantes/{id}": {
      "get": {
        "tags": [
          "Postulantes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Postulante"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulante"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Postulante"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Postulantes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Postulante"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Postulante"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Postulante"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Postulantes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Vacantes": {
      "get": {
        "tags": [
          "Vacantes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vacante"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vacante"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vacante"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vacantes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vacante"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Vacante"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Vacante"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vacante"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vacante"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vacante"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vacantes/{id}": {
      "get": {
        "tags": [
          "Vacantes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vacante"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vacante"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vacante"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vacantes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vacante"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Vacante"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Vacante"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Vacantes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Empresa": {
        "required": [
          "direccion",
          "nombre",
          "rnc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "rnc": {
            "minLength": 1,
            "type": "string"
          },
          "direccion": {
            "minLength": 1,
            "type": "string"
          },
          "sitioWeb": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Postulacion": {
        "required": [
          "postulanteId",
          "vacanteId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "vacante": {
            "$ref": "#/components/schemas/Vacante"
          },
          "postulante": {
            "$ref": "#/components/schemas/Postulante"
          },
          "postulanteId": {
            "type": "integer",
            "format": "int32"
          },
          "vacanteId": {
            "type": "integer",
            "format": "int32"
          },
          "fechaPostulacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Postulante": {
        "required": [
          "email",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "cvUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Vacante": {
        "required": [
          "areaTrabajo",
          "descripcion",
          "modalidad",
          "tipoPosicion",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "titulo": {
            "minLength": 1,
            "type": "string"
          },
          "descripcion": {
            "minLength": 1,
            "type": "string"
          },
          "tipoPosicion": {
            "minLength": 1,
            "type": "string"
          },
          "areaTrabajo": {
            "minLength": 1,
            "type": "string"
          },
          "modalidad": {
            "minLength": 1,
            "type": "string"
          },
          "empresa": {
            "$ref": "#/components/schemas/Empresa"
          },
          "salario": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "fechaPublicacion": {
            "type": "string",
            "format": "date-time"
          },
          "fechaExpiracion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}